+++ ios/Classes/PdftronFlutterPlugin.h
@@ -119,6 +119,8 @@ static NSString * const PTFormCreateRadioFieldToolKey = @"FormCreateRadioField";
 static NSString * const PTFormCreateComboBoxFieldToolKey = @"FormCreateComboBoxField";
 static NSString * const PTFormCreateListBoxFieldToolKey = @"FormCreateListBoxField";
 static NSString * const PTPencilKitDrawingToolKey = @"PencilKitDrawing";
+static NSString * const PTAnnotationSmartPenToolKey = @"AnnotationSmartPen";
+static NSString * const PTPanToolKey = @"Pan";
 
 // button
 static NSString * const PTStickyToolButtonKey = @"stickyToolButton";