+++ lib/src/document_view.dart
@@ -90,6 +90,12 @@ class DocumentViewController {
         Functions.importAnnotations, <String, dynamic>{Parameters.xfdf: xfdf});
   }
 
+  /// Merges the given XFDF annotation string to the current document.
+  Future<void> mergeAnnotations(String xfdf) {
+    return _channel.invokeMethod(
+        Functions.mergeAnnotations, <String, dynamic>{Parameters.xfdf: xfdf});
+  }
+
   /// Exports the specified annotations in the current document as a XFDF annotation string.
   ///
   /// ```dart
@@ -401,7 +407,7 @@ class DocumentViewController {
   /// [path] specifies the file name of the image resource. The button will use
   /// the specified icon if [Config.showLeadingNavButton], which is true by
   /// default, is true in the config. To add an image file to your application,
-  /// please follow the steps in the [wiki page](https://github.com/PDFTron/pdftron-flutter/wiki/Adding-an-Image-Resource-To-Your-Application).
+  /// please follow the steps in the [wiki page](https://github.com/ApryseSDK/pdftron-flutter/wiki/Adding-an-Image-Resource-To-Your-Application).
   Future<void> setLeadingNavButtonIcon(String path) {
     return _channel.invokeMethod(Functions.setLeadingNavButtonIcon,
         <String, dynamic>{Parameters.leadingNavButtonIcon: path});