+++ CHANGELOG.md
@@ -1,3 +1,70 @@
+# 1.0.0-40 - February 19, 2025
+- Update Android to v11.3.0
+
+# 1.0.0-39 - January 6, 2025
+- Update Android to v11.2.0
+
+# 1.0.0-37 - November 25, 2024
+- Update Android to v11.1.0
+
+# 1.0.0-36 - October 28, 2024
+- Update Android to v11.0.0
+- Update Android targetSDK to 34
+
+# 1.0.0-33 - September 4, 2024
+- Update Android to v10.12.0
+
+# 1.0.0-32 - July 24, 2024
+- Update Android to v10.11.0
+
+# 1.0.0-31 - May 27, 2024
+- Fix bug in `startAnnotationToolbarItemPressedListener`
+
+# 1.0.0-30 - May 1, 2024
+- Update Android to v10.9.0
+
+# 1.0.0-29 - March 20, 2024
+- Update Android to v10.8.0
+
+# 1.0.0-27 - February 7, 2024
+- Update Android to v10.7.0
+
+# 1.0.0-24 - December 13, 2023
+- expose Pan tool
+
+# 1.0.0-23 - December 6, 2023
+- Update Android to v10.6.0
+
+# 1.0.0-22 - October 25, 2023
+- Update Android to v10.5.0
+
+# 1.0.0-21 - September 21, 2023
+- Fix bug in mergeAnnotations()
+
+# 1.0.0-20 - September 13, 2023
+- Update Android to v10.4.0
+
+# 1.0.0-18 - August 23, 2023
+- mergeAnnotations()
+
+# 1.0.0-17 - August 8, 2023
+- added eraser to PTToolKey (iOS)
+  
+# 1.0.0-16 - August 2, 2023
+- Update Android to v10.3.0
+
+# 1.0.0-15 - July 4, 2023
+- Fix bug in Android `saveDocument`
+
+# 1.0.0-14 - June 21, 2023
+- Update Android to v10.2.0
+
+# 1.0.0-13 - May 10, 2023
+- Update Android to v10.1.0
+
+# 1.0.0-9 - Mar 27, 2023
+- Update Android to v10.0.0
+
 # 1.0.0-8 - Feb 17, 2023
 - Update Android to v9.5.0
 
@@ -76,7 +143,7 @@
 
 # 1.0.0-beta.3 - November 8, 2021
 
-- Updated the Android plugin and widget ([#128](https://github.com/PDFTron/pdftron-flutter/issues/128)):
+- Updated the Android plugin and widget ([#128](https://github.com/ApryseSDK/pdftron-flutter/issues/128)):
   - The plugin now supports [Flutter's new embedding engine](https://flutter.dev/docs/development/packages-and-plugins/plugin-api-migration).
   - The widget now uses [hybrid composition](https://flutter.dev/docs/development/platform-integration/platform-views?tab=ios-platform-views-objective-c-tab#hybrid-composition). This update fixes issues such as: https://github.com/flutter/flutter/issues/58273
     and in regards to stability, places the widget on parity with the plugin.